Generify TabDeckEditor
