+++ cockatrice/src/client/ui/widgets/visual_deck_storage/deck_preview/deck_preview_deck_tags_display_widget.h
@@ -2,7 +2,6 @@
 #define DECK_PREVIEW_DECK_TAGS_DISPLAY_WIDGET_H
 
 #include "../../../../../deck/deck_loader.h"
-#include "../../../../tabs/tab_deck_editor.h"
 #include "deck_preview_widget.h"
 
 #include <QWidget>
